refactor(frontend): extract useTheme and useAddressedSuggestions custom hooks (#189)
